Refactor part 2
