+++ package.json
@@ -1,11 +1,8 @@
 {
   "name": "keditor",
   "version": "2.0.0",
-  "description": "KEditor is a JQuery plugin which provides a content editor with drag and drop snippets. Uses CK Editor for inline editing of blocks.",
+  "description": "KEditor is a jQuery plugin which provides a content editor with drag and drop snippets. Uses CK Editor for inline editing of blocks.",
   "main": "gulpfile.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
   "repository": {
     "type": "git",
     "url": "git+https://github.com/Kademi/keditor.git"
@@ -16,7 +13,7 @@
     "content editor"
   ],
   "author": "Kademi (http://kademi.co)",
-  "license": "For licensing, see LICENSE.md",
+  "license": "MIT",
   "bugs": {
     "url": "https://github.com/Kademi/keditor/issues"
   },
