+++ src/keditor/constants/snippetType.js
@@ -1,0 +1,7 @@
+const SNIPPET_TYPE = {
+    ALL: 0,
+    CONTAINER: 1,
+    COMPONENT: 2
+};
+
+export default SNIPPET_TYPE;
