+++ src/keditor/constants/toolbarType.js
@@ -1,0 +1,10 @@
+const TOOLBAR_TYPE = {
+    CONTENT_AREA: 0,
+    CONTAINER: 1,
+    SUB_CONTAINER: 2,
+    CONTAINER_CONTENT: 3,
+    SUB_CONTAINER_CONTENT: 4,
+    COMPONENT: 5
+};
+
+export default TOOLBAR_TYPE;
