+++ src/keditor/utils/error.js
@@ -1,0 +1,5 @@
+const error = (message) => {
+    throw new Error(`[ KEditor ] ${message}`);
+};
+
+export default error;
